feat: extract event form
