+++ src/routes/dashboard/admin/events/eventsHelper.ts
@@ -61,86 +61,3 @@ export async function loadSpeaker(fetch: typeof globalThis.fetch, eventId: numbe
         `Unerwartete Daten für alle Speaker im Event ${eventId}.`,
     );
 }
-
-export function validateData(
-    data: SetAdminEvent,
-    allSpeaker: SetAllAdminEventSpeaker,
-    allEvents: DashboardAllEvents,
-): string[] {
-    const errorList: string[] = [];
-
-    // name
-    for (const event of allEvents) {
-        if (event.id == data.id) {
-            continue;
-        }
-        if (event.title.trim() === data.title.trim()) {
-            errorList.push(`Event mit dem Titel ${data.title} existiert bereits.`);
-        }
-    }
-
-    // text
-    if (data.title.trim().length === 0) {
-        errorList.push('Das Feld "Titel" muss angegeben werden.');
-    }
-    if (data.subtitle.trim().length === 0) {
-        errorList.push('Das Feld "Untertitel" muss angegeben werden.');
-    }
-    if (data.description_headline.trim().length === 0) {
-        errorList.push('Das Feld "Überschrift Beschreibung" muss angegeben werden.');
-    }
-    if (data.description.trim().length === 0) {
-        errorList.push('Das Feld "Beschreibung" muss angegeben werden.');
-    }
-
-    // dates
-    if (!isBeforeOrSameDatesString(data.start_date, data.end_date)) {
-        errorList.push('Das Start-Datum liegt nach dem End-Datum.');
-    }
-
-    if (data.publish_date && data.schedule_visible_from) {
-        if (!isBeforeOrSameDatesString(data.publish_date, data.schedule_visible_from)) {
-            errorList.push('Der Ablaufplan ist vor dem Event sichtbar.');
-        }
-    }
-    if (data.publish_date) {
-        if (!isBeforeOrSameDatesString(data.publish_date, data.start_date)) {
-            errorList.push('Das Event wird erst nach dem Event-Start veröffentlicht.');
-        }
-    }
-    if (data.schedule_visible_from) {
-        if (!isBeforeOrSameDatesString(data.schedule_visible_from, data.start_date)) {
-            errorList.push('Der Ablaufplan ist erst nach dem Event-Start sichtbar.');
-        }
-    }
-
-    if (data.call_for_papers_start && data.call_for_papers_end) {
-        if (!isBeforeOrSameDatesString(data.call_for_papers_start, data.call_for_papers_end)) {
-            errorList.push('Das Anmeldeende liegt vor dem Anmeldestart.');
-        }
-    }
-
-
-    // url
-    const urlScheme = z.string().url().nullable();
-    if (!urlScheme.safeParse(data.discord_url).success) {
-        errorList.push('Die Discord-URL ist nicht valide.');
-    }
-    if (!urlScheme.safeParse(data.presskit_url).success) {
-        errorList.push('Die Presskit-URL ist nicht valide.');
-    }
-    if (!urlScheme.safeParse(data.twitch_url).success) {
-        errorList.push('Die Twitch-URL ist nicht valide.');
-    }
-
-    // speaker
-    for (const speaker of allSpeaker) {
-        if (speaker.visible_from) {
-            if (!isBeforeOrSameDatesString(speaker.visible_from, data.start_date)) {
-                errorList.push(`Speaker ${speaker.name} ist erst nach dem Event-Start sichtbar.`);
-            }
-        }
-    }
-
-    return errorList;
-}
